TextConfig: config order determined when adding blocks; sub-blocks do not need unique names
TextConfig: config order determined when adding blocks; sub-blocks do not need to be unique

Previously, the order tht blocks would be configured was determined by the order that
the blocks appeared in the yaml file. This could lead to some strange issues when the
order that the blocks are configured in matters. The order that the blocks are
configured now is by default the order that the blocks are added to the TextConfig
class instance but, this can be changed by specifying a `pos` arguement that
corresponds to the name of a previously added algorithm.

The sub-blocks must now be associated with a super-block using the `superBlocks` arguement
when adding the config block. This can be a list or a single string. Since the sub-blocks
are now asscoiated with the super-blcoks, the same name in the yaml file can
correspond to different config blocks depending on the context.