DerivationFrameworkTileCal: Fix modifying locked data in TCAL1 derivation (ADCSUPPORT-5582)
TCAL1 derivation production tools have been modified to use write decoration handles
per muon container and not per each muon in this container.
Since when the handles are deleted the corresponding decorations are locked
and this produces issue when several muons are decorated in the same container.

In the same time functionality to save only the part of information from Tile cells
has been removed also to simplify the  code and since all information is saved always any way.