Set the seed redundency check for the FS tracking
Set the seed redundency check for the FS tracking

In order to speed up the tracking we can remove redundent seeds before the
full pattern recognition. This provides a significant saving in CPU with only
a trivial reduction in the efficiency. The change has been valiudated in a
dedicated trigger reprocessing. It will likely change the trigger counts.
Consequently the reference will be updated.