add a first version of nominal-only overlap removal
The way this works it (should) turn off all but the nominal for the main overlap removal algorithm, and then schedules extra algorithms to copy over the decision from the nominal to all shallow copies, so that it can be used for preselections in other algorithms (if desired).

I did test that it runs and defined some test cases, but I didn't test at all that it does anything meaningful or correct.  I leave that to physics validation.