MetaReader: Print type for metadata items of POD type
MetaReader: Print object title for metadata item of POD type

In the `metadata_items` section of `meta-reader` we list the content of metadata as a pairs of `branch.GetName():branch.GetClassName()` (or some custom name for some metadata types). However `GetClassName()` seems to return an empty string for simple types. For such cases this MR adds a fallback to print `branch.GetTitle()`, which returns the `name/type`.

Practically, so far we have e.g.:

```
    ├── metadata_items    :
...
    │   ├── FileMetaDataAux.  : xAOD::FileMetaDataAuxInfo_v1
    │   ├── FileMetaDataAuxDyn.amiTag: string
    │   ├── FileMetaDataAuxDyn.beamEnergy:
...
    │   ├── FileMetaDataAuxDyn.mcProcID:
...

```

With this MR:

```
    ├── metadata_items    :
...
    │   ├── FileMetaDataAux.  : xAOD::FileMetaDataAuxInfo_v1
    │   ├── FileMetaDataAuxDyn.amiTag: string
    │   ├── FileMetaDataAuxDyn.beamEnergy: beamEnergy/F
...
    │   ├── FileMetaDataAuxDyn.mcProcID: mcProcID/F
...
```