introduce an option for enabling/disabling output branches
A fairly long-standing feature request that allows to enable and disable
individual branches in the output n-tuple.  As a bit of a silly example
of how this can work:
```
configSeq.setOptionValue ('.commands', ['disable jet_ftag.*', 'enable jet_ftag_legacy_.*'])
```

The basic idea is that every branch gets/has a flag indicating whether
it is enabled in the output or not.  This allows to disable a big chunk
of them and then re-enable a subset, but it also allows to add some
expert-level branches that are disabled by default.