Added a script to convert LHE files produced by Superchic to a Tauola-compatible format
Added a script to convert LHE files produced by Superchic to a Tauola-compatible format

The script is based on LheConverter.py in the same directory.
The only change with respect to LheConverter.py is to change the PID of the incoming photons
from 22 to +/- 11 i.e. electrons. The photon with pz > 0 is assigned PID = 11. The photon
with pz < 0 is assigned PID = -11