Avoiding use of empty decorations in OverlapRemovalGenUseAlg
Avoiding use of empty decorations in OverlapRemovalGenUseAlg

As in the subject. The issue was reported on PAT help and offline SW
help. When the decorations are set to an empty string, the code crashes:

OverlapRemovalGenUseAlg ERROR SG::ExcBadVarName: Bad name for auxiliary
variable `'. Variable and class names must consist of only digits,
letters, and underscores, and cannot start with a digit. Variable names
may not be empty, but class names may be.

The functionality, however, is meant to be there to allow the setting of
empty decoration names in order to skip certain operations in the
algorithm.

This patch should just move things around a bit so that if empty
decorations are in place, we don't get a crash and get the expected
behavior.