RootStorageSvc: Ignore EINVAL in final file flush.
RootStorageSvc: Ignore EINVAL in final file flush.

In RootDatabase::close(), we call TFile::Write to flush the file before
it is closed.  In doing so, we watch to see if errno gets set to something;
if it does, we complain.
However, TFile::Write can trigger ROOT to search for libraries.
This can result in errno being set to ENOENT.  Due to this, ENOENT
was already being ignored here.
But on aarch64 hosts, we also see ROOT calling readlinkat() while
searching for libraries, on paths that are not symbolic links.
This can result in errno being set to EINVAL.  So ignore EINVAL as well.

Fixes test failures on aarch64 in this package as well as APRTests
and IOVDbTestAlg.