TrkVertexFitters: Fix precision issue in test.
TrkVertexFitters: Fix precision issue in test.

The AdaptiveMultiVertexFitter test was failing on arm.
This is because the input covariance matrix for the vertex positions
was set so large that we were losing all precision in the matrix
inversion in KalmanVertexUpdator::positionUpdate.

Made the input more reasonable, and now we get consistent results
between arm and x86.