Add a histogram for the roi eta
Add a histogram for the roi eta

Frequently the trigger tracking results appear to have changed, but
often this is due not to the tracking itself, but the upstream processing
which provides a different distribution ofRois.

This adds a histogram to keep a record of the input Roi eta postion to
detmine whether the input Rois are indeed the same as the reference.