tauRec: early stopping of EleRM tau reconstruction
Hello,

This MR is implementing an early stopping condition for the electron-subtracted ("EleRM") tau reconstruction.
If no suitable (medium likelihood) electron is found in the event, there is no reason to run the EleRM tau reconstruction.
The easiest way to "abort" the EleRM reconstruction is to produce an empty electron-subtracted topocluster container in the first EleRM algorithm (TauElecSubtractAlg): no cluster means no seed jet to be reconstructed, therefore no EleRM tau.
Currently, there is an early stopping condition, but it happens later in the chain. Here, we move it as early as it can be.

The CPU saving seems decent. Diff-root finds no change in the AOD output, tested on 500 ttbar events.

Cheers,
Bertrand