Correcting for the factor in the outputs of the DIPZ regression in Athena
Correcting the for the factor in the outputs of the DIPZ regression in Athena

The original current DIPZ output from Athena has a factor of "50" in the outputs that we want to get rid of. Thus we derived another neural network with an extral layer that does such change and validated it. We defined the output of the network with a new timestamp. Here, we modified the files that use DIPZ with the new timestamped variable name, and also pointed athena to the location of the new json file.

Testing of these changes is in progress...