Adding keywords to evgen metadata
Keywords in event generation specify various things about the process
being generated - what physics it might represent, what particles are
being generated, and so on.

These keywords are extremely handy downstream for auto configuration.
See for example the discussion in !67455. Similarly, one can configure
SUSY helpers for all SUSY samples based on keyword, or Higgs helpers for
all Higgs samples, and so on.

This adds the keywords to the EVNT metadata, so that they can be
transferred along and accessed downstream. It'll be a long time before
all our evgen have these keywords, but better to start the process now.