Provide an example of setting ExtraDataForDynamicConsumers in a DAOD job with string parsing
Provide an example of setting ExtraDataForDynamicConsumers in a DAOD job with string parsing

This amends the TEST2 example DAOD format to demonstrate how to set ExtraDataForDynamicConsumers in a job that uses ExpressionEvaluation. This is necessary when the relevant containers being accessed by ExpressionEvaluation have not been read in vias ReadHandles. This applies to skims of PHYS(LITE) and some HI DAODs, for example. It doesn't apply to PHYS(LITE) and many other DAOD formats due to the scheduling of common augmentations, which have already read in the relevant containers.

The TEST2 format is just an example and is never used in production so this MR will have no impact on any data products. The relevant lines will be used as a template in other formats.