first test of using LeptonTaggers in MUON5
Convert r21 decorators to const Accessor for vertex fits

Previously, secondary vertex decoration was done using a syntax like

(*decorator)(*vertex) = value;

This has been changed in two places to use a const Accessor, as this
makes the logic of the decoration more obvious and might avoid some
subtle bugs with decoration. This was an attempt to resolve a warning

WARNING saveSecondaryVertices - missing "SecondaryVertexIndex" variable

but this change does not seem to fix the problem.