Adding barcodes back into the GeneratorObjects and GeneratorObjectsTPCnv unit tests.
Adding barcodes back into the GeneratorObjects and GeneratorObjectsTPCnv unit tests.

This will become important as the barcode migration (ATLASSIM-6906) continues and we need to
confirm that legacy persistent formats are treated correctly.

NB As the migration progesses barcode support will be dropped from `McEventCollection_p7` and
a barcode-independent `HepMcParticleLink_p3` will be introduced.