Restore mergeIteratively2TrackVtxs in LeptonTaggers
Make VarHolder scope local

This clears most of the "variable not decorated" warnings. The VarHolder
object in decorateElec() was originally locally scoped in r21. At some
point during the port, it became a class member. This prevented the
variables from being decorated correctly to each electron. Transitioning
back to a locally-scoped variable resolves the problem.