AthContainers+AthContainersInterfaces: Fix return value of AuxTypeVector::resize for null container
AthContainers: Fix return value of AuxTypeVector::resize for null container

AuxTypeVector::resize is supposed to return false if iterators were
invalidated.  We were doing this by testing if the allocated block
moves during the resize.

This was being used to know when caches must be invalidated.

However, if we resize an empty container that has reserved memory, then
the the allocated block may not move.  But in that case we were 
sometimes failing to invalidate a dependent cache.

Change so that we explicitly return false when increasing the size
(and clarify documentation).