DerivationFrameworkMCTruth: keep all daughters of TruthTausWithDecayParticles
DerivationFrameworkMCTruth: keep all daughters of TruthTausWithDecayParticles

Hello,

This MR changes the number of generations of daughter particles retained when building the TruthTausWithDecayParticles/TruthTausWithDecayVertices collections.
It was previously set to 1, but this can lead to pathological situations, e.g. when the truth tau decays to kaons: if we don't know what the kaon decays to, we cannot properly compute the truth visible tau momentum.
When set to -1, we retain all generations, and this seems to fix the problem.

 A priori, we can now recompute the various decorations of TruthTaus using TruthTausWithDecayParticles/TruthTausWithDecayVertices, such as `pt/eta/phi/m_vis`, the number of neutral/charged pions, ...
We will need to rework a bit BuildTruthTaus to do this.
The only variables we don't seem to get is `polarizationPhi` and `polarizationTheta` decorations, they exist for TruthTaus but not TruthTausWithDecayParticles, however these seem to always be 0.
We still need to check we can rerun photon dressing for TruthTausWithDecayParticles.

Marking as urgent as we'd need this for the upcoming DAOD_PHYS production that is imminent.

Cheers,
Bertrand