Replace ConfigFactory used in block config with backend for text config so both are consistent.
Default blocks are added by default. PRW parameters are obtained from ConfigAccumulator.

Add autoconfigFlags as an attribute to the ConfigAccumulator so information can be
obtained in other blocks.

Update the PileupReweighting block so that campaign, input files, and useDefaultConfig
can be obtained from autoconfigFlags. The wrapper function that handeled this before has
been removed.

addDefaultAlgs function is now part of the ConfigFactory class and is called
automatically when initializing an object instance. Default parameters are now
obtained from ConfigAccumulator except for noSystematics flag so that this can
be changed from the command line if not specified directly.