AthContainers: Fix BaseInfo initialization.
AthContainers: Fix BaseInfo initialization.

When we load a library containing a DataVector that does not have
a DATAVECTOR_BASE declaration, the BaseInfo for the DataVector class
may not get created until the class is actually used.  Add a reference
to the BaseInfo instance to the DVLInfo ctor to force it to be created
earlier.

See ATEAM-674.